추가: user 데이터 및 시딩
