+++ prisma/seed.js
@@ -1,18 +1,28 @@
 import { PrismaClient } from '@prisma/client';
-import { mockArticles, mockProducts } from '../src/utils/mock.js';
+import { mockUsers, mockArticles, mockProducts } from './mock.js';
 
 const prisma = new PrismaClient();
 
 async function main() {
+  await prisma.productLike.deleteMany();
+  await prisma.articleLike.deleteMany();
   await prisma.comment.deleteMany();
   await prisma.article.deleteMany();
   await prisma.product.deleteMany();
+  await prisma.user.deleteMany();
+
+  for (const userData of mockUsers) {
+    await prisma.user.create({
+      data: userData,
+    });
+  }
 
   for (const articleData of mockArticles) {
     await prisma.article.create({
       data: {
         title: articleData.title,
         content: articleData.content,
+        userId: articleData.userId,
         comments: {
           create: articleData.comments,
         },
@@ -27,6 +37,7 @@ async function main() {
         description: productData.description,
         price: productData.price,
         tags: productData.tags,
+        userId: productData.userId,
         comments: {
           create: productData.comments,
         },
