--- src/utils/mock.js
@@ -1,78 +1,0 @@
-export const mockArticles = [
-  {
-    title: '프로그래밍 잘 하는 방법',
-    content: '코드를 많이 써보면 된답니다',
-    comments: [
-      { content: '와, 정말 유용한 정보네요!' },
-      { content: '다 아는데 안 하는 거 뿐이죠 ㅋㅋ..' },
-    ],
-  },
-  {
-    title: '저는 미션 하면서 노래를 틀어놓고 합니다',
-    content:
-      '원래 케이팝을 들었는데 오늘은 팝송을 틀어봤어요 그시절 팝송 이런 플레이리스트 추천합니다',
-    comments: [{ content: '오 신나게 코딩하시네요' }],
-  },
-  {
-    title: '미션하고 프로젝트 하려고 합니다',
-    content: '빠듯하네요',
-    comments: [],
-  },
-  {
-    title: 'Superstruct로 유효성 검사',
-    content:
-      'object, string, number 등 스키마를 정의하고 validate 함수나 assert 함수로 검증할 수 있습니다.',
-    comments: [{ content: 'StructError... 이름이 어렵네요.' }],
-  },
-  {
-    title: '오늘의 점심 메뉴 추천',
-    content: '날씨가 쌀쌀하니 따뜻한 국밥이나 칼국수는 어떠신가요? 든든하게 먹고 오후도 힘내세요!',
-    comments: [
-      { content: '마침 배고팠는데, 국밥 먹으러 갑니다!' },
-      { content: '저는 칼국수요.' },
-      { content: '제육도 좋아요.' },
-    ],
-  },
-];
-
-export const mockProducts = [
-  {
-    name: '고성능 기계식 키보드',
-    description: '청축(Clicky) 타입의 타건감이 뛰어난 기계식 키보드입니다. RGB 백라이트 지원.',
-    price: 129000,
-    tags: ['전자기기', '키보드', '게이밍'],
-    comments: [
-      { content: '타건감 정말 좋네요. 만족합니다.' },
-      { content: '밤에 쓰기엔 조금 시끄러워요.' },
-    ],
-  },
-  {
-    name: '4K UHD 모니터 27인치',
-    description: 'IPS 패널을 탑재한 27인치 4K 모니터. HDR10을 지원하여 생생한 화질을 제공합니다.',
-    price: 345000,
-    tags: ['전자기기', '모니터', '4K'],
-    comments: [{ content: '화질이 정말 선명해요. 넷플릭스 볼 맛 납니다.' }],
-  },
-  {
-    name: '무소음 무선 마우스',
-    description:
-      '도서관이나 조용한 사무실에서도 사용하기 좋은 무소음 클릭 마우스입니다. USB 리시버 타입.',
-    price: 25000,
-    tags: ['전자기기', '마우스', '사무용품'],
-    comments: [],
-  },
-  {
-    name: '프리미엄 원두 커피 500g',
-    description: '브라질 세하도 원두를 미디엄 로스팅하여 고소하고 부드러운 맛이 특징입니다.',
-    price: 18000,
-    tags: ['식품', '커피', '원두'],
-    comments: [{ content: '향이 정말 좋아요. 재구매 의사 있습니다.' }],
-  },
-  {
-    name: '대용량 보조 배터리 20000mAh',
-    description: 'USB-C PD 충전을 지원하는 대용량 보조 배터리. 스마트폰, 태블릿 동시 충전 가능.',
-    price: 38000,
-    tags: ['전자기기', '모바일', '배터리'],
-    comments: [],
-  },
-];
